Added ability to load in saved bugs
