+++ crates/genesis_lib/src/behaviour/laying.rs
@@ -1,22 +1,16 @@
 use bevy::{
-    ecs::system::EntityCommands,
-    prelude::{
-        default, AssetServer, Bundle, Color, Commands, Entity, Handle, Image, Query, Res, ResMut,
-        Transform, Vec2, Vec3, With, Without,
-    },
-    sprite::{Sprite, SpriteBundle},
+    prelude::{AssetServer, Commands, Entity, Query, Res, ResMut, Transform, Vec3, With, Without},
+    sprite::Sprite,
 };
-use bevy_rapier2d::prelude::{ActiveEvents, Collider, Damping, RigidBody, Velocity};
-use components::SizeMultiplier;
 use genesis_attributes as attributes;
 use genesis_components as components;
-use genesis_components::{body, eat, grow, lay::*, mind, see, time, BurntEnergy};
+use genesis_components::{body, lay::*, mind};
 use genesis_config as config;
 use genesis_ecosystem as ecosystem;
 use genesis_newtype::Probability;
 use genesis_spawners::Spawners;
 
-use crate::{setup::MindThresholds, statistics};
+use crate::{setup::MindThresholds, spawning, statistics};
 
 type LayerTest<'a> = (Entity, &'a mind::MindOutput);
 
@@ -78,7 +72,7 @@ pub fn lay_egg_system(
         let energy = vitality.take_energy(egg_energy);
         let location = egg_position(transform);
         eggs_laid.0 += 1;
-        let egg_entity = spawn_egg(
+        let egg_entity = spawning::spawn_egg(
             &mut commands,
             &asset_server,
             energy,
@@ -102,131 +96,6 @@ fn egg_position(parent_transform: &Transform) -> Vec3 {
     egg_pos.x += separation * s;
 
     egg_pos
-}
-
-type BugParts<'a> = (
-    attributes::Genome,
-    mind::Mind,
-    &'a Color,
-    &'a attributes::HatchSize,
-    &'a attributes::MaxSize,
-);
-
-fn spawn_bug(
-    asset_server: &Res<AssetServer>,
-    energy: ecosystem::Energy,
-    bug_parts: BugParts,
-    mut hatching_entity: EntityCommands,
-) -> ecosystem::Energy {
-    let (bug_body, mind, color, hatch_size, max_size) = bug_parts;
-    let mind_bundle = mind::MindBundle::new(&mind);
-
-    let original_color = body::OriginalColor(mind.color());
-    // Allows selected eggs to remain selected on hatching
-    let current_color = if *color == Color::RED {
-        *color
-    } else {
-        original_color.0
-    };
-
-    let size = body::Size::new(**hatch_size, **max_size);
-    let (vitality, leftover_energy) = body::Vitality::new(size, energy);
-
-    let sprite_image: Handle<Image> = asset_server.load("sprite.png");
-    let sprite = Sprite {
-        custom_size: Some(vitality.size().sprite()),
-        color: current_color,
-        ..default()
-    };
-
-    hatching_entity
-        .insert(sprite_image)
-        .insert(sprite)
-        .insert(ActiveEvents::COLLISION_EVENTS)
-        .insert(vitality.size().collider())
-        .insert(SizeMultiplier::new(vitality.size().current_size()))
-        .insert(components::Juvenile)
-        .insert(original_color)
-        .insert(bug_body)
-        .insert(vitality)
-        .insert(mind_bundle)
-        .insert(see::Vision::new())
-        .insert(time::Age::default())
-        .insert(time::Heart::new())
-        .insert(time::InternalTimer::new())
-        .insert(components::MovementSum::new())
-        .insert(components::ThinkingSum::new())
-        .insert(eat::EatingSum::new())
-        .insert(grow::GrowingSum::new())
-        .insert(grow::SizeSum::new())
-        .insert(eat::EnergyConsumed(0))
-        .insert(EggsLaid(0));
-
-    leftover_energy
-}
-
-#[derive(Bundle)]
-struct EggBundle {
-    pub egg: components::Egg,
-    pub egg_energy: ecosystem::EggEnergy,
-    pub sprite: Sprite,
-    pub handle: Handle<Image>,
-    pub original_color: body::OriginalColor,
-    pub collider: Collider,
-    pub age: time::Age,
-}
-
-fn spawn_egg(
-    commands: &mut Commands,
-    asset_server: &Res<AssetServer>,
-    energy: ecosystem::Energy,
-    location: Vec3,
-    genome: attributes::Genome,
-    mind: mind::Mind,
-    generation: components::Generation,
-    parent_id: Option<Entity>,
-) -> Entity {
-    let size = 16.0;
-
-    let attribute_bundle = attributes::AttributeBundle::new(&genome);
-    let original_color = body::OriginalColor(Color::WHITE);
-    let sprite = SpriteBundle {
-        texture: asset_server.load("egg.png"),
-        sprite: Sprite {
-            custom_size: Some(Vec2::new(size, size)),
-            color: original_color.0,
-            ..default()
-        },
-        transform: Transform::from_translation(location),
-        ..default()
-    };
-
-    let mut egg_entity = commands.spawn(sprite);
-    let entity = egg_entity.id();
-
-    egg_entity
-        .insert(RigidBody::Dynamic)
-        .insert(Damping {
-            linear_damping: 1.0,
-            angular_damping: 1.0,
-        })
-        .insert(Velocity::zero())
-        .insert(Collider::ball(size / 2.0))
-        .insert(components::Egg)
-        .insert(attribute_bundle)
-        .insert(ecosystem::EggEnergy(energy))
-        .insert(original_color)
-        .insert(genome)
-        .insert(components::Relations::new(
-            (entity, mind.color()),
-            parent_id,
-        ))
-        .insert(mind)
-        .insert(time::Age::default())
-        .insert(generation)
-        .insert(BurntEnergy::new());
-
-    entity
 }
 
 pub fn spawn_egg_system(
@@ -255,7 +124,7 @@ pub fn spawn_egg_system(
                 .mutate(&mut rng, &Probability::new(1.0).unwrap(), &mind_thresholds)
                 .into();
         }
-        spawn_egg(
+        spawning::spawn_egg(
             &mut commands,
             &asset_server,
             energy,
@@ -287,9 +156,9 @@ pub fn hatch_egg_system(
     for (entity, mut egg_energy, mind, genome, sprite, hatch_size, max_size) in
         hatch_query.iter_mut()
     {
-        commands.entity(entity).remove::<EggBundle>();
+        commands.entity(entity).remove::<spawning::EggBundle>();
         let hatching_entity = commands.entity(entity);
-        let leftover_energy = spawn_bug(
+        let leftover_energy = spawning::spawn_bug(
             &asset_server,
             egg_energy.move_all_energy(),
             (
